[malloy core] Fix Broken Table Schema Query (#16)
TLDR: SQL query for table schema had a small but deadly error resulting in a crash.

Today, Malloy does not support any variant types. This means that arbitrary json objects need to have a fixed schema. The code as it was written ignores situations where an object might have an int or a string type. HOWEVER, one important thing the current code does not account for is having multiple collection types potentially. 

Lets take an example:
if "abc" is a json object that contains the keys "name" and "value" and "value" can sometimes be string and sometimes int, malloy for snowflake will just ignore the untypable field (I think, but based on comments in the code this is true).

BUT if "abc" has a key "stuff" which can be EITHER a json or an array object, snowflake will skip processing this field but might still try to process "stuff".value or "stuff"[0]. 

This results in a crash.

The fix in this PR is to ensure that when we are querying for the table schema, we drop all descendants of ambiguous columns along with that column. This will ensure the rest of the code does not crash and that we can actively read information from such fully variant things.